Organize automation scripts and rename for clarity
